sink: move to the `channel` module
The `common` package is an antipattern.
