+++ pkg/peer/peer.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/matrix-org/waterfall/pkg/common"
+	"github.com/matrix-org/waterfall/pkg/channel"
 	"github.com/matrix-org/waterfall/pkg/peer/state"
 	"github.com/matrix-org/waterfall/pkg/webrtc_ext"
 	"github.com/pion/rtcp"
@@ -31,7 +31,7 @@ var (
 type Peer[ID comparable] struct {
 	logger         *logrus.Entry
 	peerConnection *webrtc.PeerConnection
-	sink           *common.SinkWithSender[ID, MessageContent]
+	sink           *channel.SinkWithSender[ID, MessageContent]
 	state          *state.PeerState
 }
 
@@ -39,7 +39,7 @@ type Peer[ID comparable] struct {
 func NewPeer[ID comparable](
 	connectionFactory *webrtc_ext.PeerConnectionFactory,
 	sdpOffer string,
-	sink *common.SinkWithSender[ID, MessageContent],
+	sink *channel.SinkWithSender[ID, MessageContent],
 	logger *logrus.Entry,
 ) (*Peer[ID], *webrtc.SessionDescription, error) {
 	peerConnection, err := connectionFactory.CreatePeerConnection()
